Rename the schema package to handle versioning resets
