+++ packages/db-schema/package.json
@@ -1,6 +1,6 @@
 {
-	"name": "@db-studio/schema",
-	"version": "9.26.4",
+	"name": "@db-studio/db-schema",
+	"version": "1.0.0",
 	"description": "Utility for extracting information about existing DB schema",
 	"type": "module",
 	"keywords": [
