+++ packages/db-schema/src/utils/extract-type.ts
