+++ packages/db-schema/src/utils/strip-quotes.ts
